requirements.txt
python-terrier>=0.10
sentence-transformers>=2.2
ir-datasets>=0.5
pytrec-eval-terrier>=0.5
faiss-cpu>=1.7.4
pandas>=2.0
numpy>=1.24
tqdm>=4.65
streamlit>=1.30
matplotlib>=3.7
